first round of changing type of timecodes to int64_t
use matching typ for MIN/MAX macros after [25976] and [25978]
still many implicit conversions

git-svn-id: http://svn.mythtv.org/svn/trunk@26023 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/NuppelVideoRecorder.cpp
@@ -2652,7 +2652,7 @@ void NuppelVideoRecorder::FormatCC(struct cc *cc)
 }
 
 void NuppelVideoRecorder::AddTextData(unsigned char *buf, int len,
-                                      long long timecode, char /*type*/)
+                                      int64_t timecode, char /*type*/)
 {
     int act = act_text_buffer;
     if (!textbuffer[act]->freeToBuffer)
